3 hw fixes

+++ src/main/java/ok/dht/test/anikina/ConsistentHashingImpl.java
@@ -11,16 +11,16 @@ import java.util.Map;
 class ConsistentHashingImpl {
     private static final int VIRTUAL_NODES_COUNT = 3;
 
-    private final long[] hashes;
-    private final Map<Long, String> hashToShard = new HashMap<>();
+    private final int[] hashes;
+    private final Map<Integer, String> hashToShard = new HashMap<>();
 
     ConsistentHashingImpl(List<String> clusterUrls) {
-        hashes = new long[clusterUrls.size() * VIRTUAL_NODES_COUNT];
+        hashes = new int[clusterUrls.size() * VIRTUAL_NODES_COUNT];
         for (int i = 0; i < clusterUrls.size(); i++) {
             String serverUrl = clusterUrls.get(i);
             for (int j = 0; j < VIRTUAL_NODES_COUNT; j++) {
                 String virtualNode = serverUrl + "_" + j;
-                long hash = hashForKey(virtualNode);
+                int hash = hashForKey(virtualNode);
                 hashes[i * VIRTUAL_NODES_COUNT + j] = hash;
                 hashToShard.put(hash, serverUrl);
             }
@@ -28,13 +28,13 @@ class ConsistentHashingImpl {
         Arrays.sort(hashes);
     }
 
-    private long hashForKey(String key) {
+    private int hashForKey(String key) {
         byte[] keyBytes = Utf8.toBytes(key);
         return Hash.xxhash(keyBytes, 0, keyBytes.length);
     }
 
     String getShardByKey(String key) {
-        long hash = hashForKey(key);
+        int hash = hashForKey(key);
         int shardIndex = Arrays.binarySearch(hashes, hash);
         if (shardIndex < 0) {
             shardIndex = -shardIndex - 1;
